Создать шаблоны DTO #10 Вынести сущности в отдельную библиотеку, исправить согласно БД. #23
