+++ CommonLib/Classes/Admin.cs
@@ -1,10 +1,10 @@
 ﻿using WebApp_GKH.Helpers;
-namespace WebApp_GKH.Models
+namespace CommonLib.Classes
 {
 
     public class Admin
     {
-        public Guid Id { get; set; }
+        public int Id { get; set; }
         public string FirstName { get; set; }
         public string SecondName { get; set; }
         public UserType Type { get; set; }
@@ -15,7 +15,6 @@ namespace WebApp_GKH.Models
 
         public Admin(string name, string surname, string login, string password, string dateOfBirth)
         {
-            Id = Guid.NewGuid();
             FirstName = name;
             SecondName = surname;
             //Type = Id Админа; // Id UserType
@@ -26,7 +25,6 @@ namespace WebApp_GKH.Models
 
         public Admin()
         {
-            Id = Guid.NewGuid();
             //Type = Id Админа; // Id UserType
         }
     }
