+++ CommonLib/Classes/Resident.cs
@@ -1,11 +1,11 @@
 ﻿using WebApp_GKH.Helpers;
 
-namespace WebApp_GKH.Models
+namespace CommonLib.Classes
 {
     public class Resident
     {
 
-        public Guid Id { get; set; }
+        public int Id { get; set; }
         public string FirstName { get; set; }
         public string SecondName { get; set; }
         public UserType Type { get; set; }
@@ -20,7 +20,6 @@ namespace WebApp_GKH.Models
 
         public Resident(string name, string surname, string login, string password, string dateOfBirth, string email, string phone, Address address)
         {
-            Id = Guid.NewGuid();
             FirstName = name;
             SecondName = surname;
             //Type = Id Жителя; // Id UserType
@@ -35,7 +34,6 @@ namespace WebApp_GKH.Models
 
         public Resident()
         {
-            Id = Guid.NewGuid();
             //Type = Id Жителя; // Id UserType
             DateRegistration = DateTime.Now;
         }
